Renamed files to match the name of the type it holds.
