+++ src/Scientist/Internals/InMemoryObservationPublisher.cs
